feat: internal auth, gardens and users (#427)

+++ src/container/container.service.ts
@@ -1,9 +1,9 @@
-import { forwardRef, Inject, Injectable } from '@nestjs/common';
+import { forwardRef, Inject, Injectable, NotFoundException } from '@nestjs/common';
 import { InjectModel } from '@nestjs/mongoose';
-import { FilterQuery, Model } from 'mongoose';
+import { Model, PipelineStage, Types } from 'mongoose';
+import { GardenService } from '../garden/garden.service';
 import { TaskType } from '../interface';
 import { PlantInstanceService } from '../plant-instance/plant-instance.service';
-import { PlantService } from '../plant/plant.service';
 import { TaskService } from '../task/task.service';
 import { fromTaskTypeToHistoryStatus } from '../util/history.util';
 import { isNotNullish } from '../util/null.util';
@@ -18,30 +18,99 @@ export class ContainerService {
   constructor(
     @InjectModel('Container')
     private readonly containerModel: Model<ContainerDocument>,
-    @Inject(forwardRef(() => PlantService)) private plantService: PlantService,
     @Inject(forwardRef(() => TaskService)) private taskService: TaskService,
+    @Inject(forwardRef(() => GardenService)) private gardenService: GardenService,
     @Inject(forwardRef(() => PlantInstanceService)) private plantInstanceService: PlantInstanceService
   ) {}
 
-  async addContainer(containerDTO: ContainerDTO): Promise<ContainerDocument> {
-    const newContainer = await this.containerModel.create(sanitizeContainerDTO(containerDTO));
+  async addContainer(containerDTO: ContainerDTO, userId: string, gardenId: string): Promise<ContainerDocument> {
+    const garden = this.gardenService.getGarden(gardenId, userId);
+    if (!garden) {
+      throw new NotFoundException('Garden does not exist!');
+    }
+
+    const newContainer = await this.containerModel.create({
+      ...sanitizeContainerDTO(containerDTO),
+      gardenId: new Types.ObjectId(gardenId)
+    });
     return newContainer.save();
   }
 
-  async getContainer(containerId: string | null | undefined): Promise<ContainerDocument | null> {
+  async getContainers(
+    userId: string,
+    gardenId: string,
+    extraPipeline: PipelineStage[] = []
+  ): Promise<ContainerDocument[]> {
+    const containers = await this.containerModel
+      .aggregate<ContainerDocument>([
+        {
+          $lookup: {
+            from: 'gardens',
+            localField: 'gardenId',
+            foreignField: '_id',
+            as: 'garden'
+          }
+        },
+        {
+          $unwind: {
+            path: '$garden',
+            preserveNullAndEmptyArrays: true
+          }
+        },
+        {
+          $match: {
+            'garden._id': new Types.ObjectId(gardenId),
+            'garden.userId': new Types.ObjectId(userId)
+          }
+        },
+        ...extraPipeline,
+        {
+          $project: {
+            _id: 1,
+            name: 1,
+            gardenId: 1,
+            type: 1,
+            rows: 1,
+            columns: 1,
+            slots: 1,
+            startedFrom: 1,
+            archived: 1
+          }
+        }
+      ])
+      .exec();
+
+    return containers;
+  }
+
+  async getContainer(
+    containerId: string | null | undefined,
+    userId: string,
+    gardenId: string
+  ): Promise<ContainerDocument | null> {
     if (!containerId) {
       return null;
     }
 
-    return this.containerModel.findById(containerId).exec();
-  }
+    const container = await this.getContainers(userId, gardenId, [
+      {
+        $match: {
+          _id: new Types.ObjectId(containerId)
+        }
+      }
+    ]);
 
-  async getContainers(filter: FilterQuery<ContainerDocument> = {}): Promise<ContainerDocument[]> {
-    return this.containerModel.find(filter).exec();
+    if (container.length === 0) {
+      return null;
+    }
+
+    return container[0];
   }
 
   async editContainer(
     containerId: string,
+    userId: string,
+    gardenId: string,
     containerDTO: ContainerDTO,
     updateTasks: boolean
   ): Promise<ContainerDocument | null> {
@@ -50,7 +119,7 @@ export class ContainerService {
       return null;
     }
 
-    const oldContainer = await this.getContainer(containerId);
+    const oldContainer = await this.getContainer(containerId, userId, gardenId);
     if (!oldContainer) {
       return null;
     }
@@ -99,18 +168,23 @@ export class ContainerService {
     const editedContainer = await this.containerModel.findByIdAndUpdate(containerId, newContainerDTO, { new: true });
 
     if (editedContainer && updateTasks) {
-      await this.createUpdatePlantTasks(editedContainer);
+      await this.createUpdatePlantTasks(editedContainer, userId, gardenId);
     }
 
     return editedContainer;
   }
 
-  async deleteContainer(containerId: string): Promise<ContainerDocument | null> {
+  async deleteContainer(containerId: string, userId: string, gardenId: string): Promise<ContainerDocument | null> {
+    const oldContainer = await this.getContainer(containerId, userId, gardenId);
+    if (!oldContainer) {
+      return null;
+    }
+
     const result = this.containerModel.findByIdAndRemove(containerId);
 
-    const plantInstances = await this.plantInstanceService.getPlantInstancesByContainer(containerId);
+    const plantInstances = await this.plantInstanceService.getPlantInstancesByContainer(containerId, userId, gardenId);
     for (const plantInstance of plantInstances) {
-      await this.plantInstanceService.closePlantInstance(plantInstance._id);
+      await this.plantInstanceService.closePlantInstance(plantInstance._id, userId, gardenId);
     }
 
     return result;
@@ -118,22 +192,32 @@ export class ContainerService {
 
   async updateContainerTasksByType(
     containerId: string,
+    userId: string,
+    gardenId: string,
     date: string,
     type: TaskType,
     plantInstanceIds?: string[]
   ): Promise<number> {
-    const tasks = await this.taskService.getTasks({ type, completedOn: null, start: { $lte: date } });
+    const tasks = await this.taskService.findTasks(userId, gardenId, [
+      {
+        $match: {
+          type,
+          completedOn: null,
+          start: { $lte: date }
+        }
+      }
+    ]);
 
     let updatedCount = 0;
 
     for (const task of tasks) {
-      const plantInstance = await this.plantInstanceService.getPlantInstance(task.plantInstanceId);
+      const plantInstance = await this.plantInstanceService.getPlantInstance(task.plantInstanceId, userId, gardenId);
       if (
         plantInstance &&
         plantInstance.containerId === containerId &&
         (!plantInstanceIds || (plantInstance._id && plantInstanceIds.includes(plantInstance._id.toString())))
       ) {
-        const updatedTask = await this.taskService.findByIdAndUpdate(task._id, { completedOn: date });
+        const updatedTask = await this.taskService.findByIdAndUpdate(task._id, userId, gardenId, { completedOn: date });
         updatedCount++;
 
         if (task?.type === type && updatedTask?.completedOn) {
@@ -146,7 +230,7 @@ export class ContainerService {
               subSlot: plantInstance.subSlot
             }
           });
-          await this.plantInstanceService.createUpdatePlantInstanceTasks(updatedPlantInstance);
+          await this.plantInstanceService.createUpdatePlantInstanceTasks(updatedPlantInstance, userId, gardenId);
         }
       }
     }
@@ -156,18 +240,25 @@ export class ContainerService {
 
   async createUpdatePlantTasksForSlot(
     container: ContainerDocument,
+    userId: string,
+    gardenId: string,
     slot: BaseSlotDocument,
     path: string,
     slotTitle: string,
     plantId?: string
   ) {
-    const plantInstance = await this.plantInstanceService.getPlantInstance(slot.plantInstanceId);
+    const plantInstance = await this.plantInstanceService.getPlantInstance(slot.plantInstanceId, userId, gardenId);
     if (plantInstance && (!plantId || plantInstance.plant === plantId)) {
-      this.plantInstanceService.createUpdateTasks(container, plantInstance, path, slotTitle);
+      this.plantInstanceService.createUpdateTasks(userId, gardenId, container, plantInstance, path, slotTitle);
     }
   }
 
-  async createUpdatePlantTasks(container: ContainerDocument | null | undefined, plantId?: string) {
+  async createUpdatePlantTasks(
+    container: ContainerDocument | null | undefined,
+    userId: string,
+    gardenId: string,
+    plantId?: string
+  ) {
     if (!container?.slots) {
       return;
     }
@@ -178,10 +269,18 @@ export class ContainerService {
       const path = `/container/${container._id}/slot/${slotIndex}`;
       const slotTitle = getSlotTitle(+slotIndex, container.rows);
 
-      await this.createUpdatePlantTasksForSlot(container, slot, path, slotTitle, plantId);
+      await this.createUpdatePlantTasksForSlot(container, userId, gardenId, slot, path, slotTitle, plantId);
 
       if (slot.subSlot) {
-        await this.createUpdatePlantTasksForSlot(container, slot.subSlot, `${path}/sub-slot`, slotTitle, plantId);
+        await this.createUpdatePlantTasksForSlot(
+          container,
+          userId,
+          gardenId,
+          slot.subSlot,
+          `${path}/sub-slot`,
+          slotTitle,
+          plantId
+        );
       }
     }
   }
